refactoring folder name
